requirements.txt
# PDF Generation
reportlab>=4.0.0

# Data Validation & Models
pydantic>=2.0.0
pydantic[email]>=2.0.0

# GUI Framework
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# Date/Time Utilities
python-dateutil>=2.8.2

# Configuration & Environment
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0

# Testing
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Utilities
pyyaml>=6.0.0
